Use a view instead of a table to track version
A table comes with some complexity in terms of making sure that it both
exists, and has only one value in it. Either we have to create the table
forcefully, or demand the user create a table with a specific
specification for us. Using a view, on the other hand has the dual
benefit of being easy to define with a static value without creating any
dependent constraints, but also appears as part of the schema-only dump
of the database, meaning we do not have to juggle dumping the schema
version alongside the schema-only data. There might still be some
eventual benefit to having a table that tracks what migrations have been
applied and in what order for record-keeping purposes, but that should
be a separate concern to the question of "what version is the schema
supposed to be in right now".