Keep certain SET commands
Some `SET` commands (such as `check_function_bodies`) are necessary to
keep in the interest of loading dumped files properly, later. Others,
such as the timing constraints and configurations, can cause problems
loading files in different environments. This commit defines an initial
list of kept `SET` commands that will be preserved in the dumped file.